fix(person-components): rename mock
