+++ storybook/src/stories/person/avatar-snapshoots.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Canvas } from '@storybook/addon-docs';
 import { Provider, PersonAvatar, PersonAccountType } from '@equinor/fusion-react-person';
-import { createResolve } from './resolve-mock/PersonResolve';
+import { createResolve } from './resolve-mock/person-resolve-mock';
 
 <Meta title="Snapshots/Person/PersonAvatar" parameters={{ chromatic: { disableSnapshot: false } }} />
